Divided map.js to logical chunks
